Replace old styled components with adjusted ones
